added new file + added text and orginized the code more
