+++ src/index.html
@@ -85,5 +85,6 @@
     </section>
     <script type="module" src="showUserLocation.js" defer></script>
     <script type="module" src="addMarkersLogic.js" defer></script>
+    <script type="module" src="showMarkersLogic.js" defer></script>
   </body>
 </html>
