--- src/script.ts
@@ -1,1 +1,0 @@
-console.log("hallo world");
