+++ src/showUserLocation.js
@@ -1,3 +1,5 @@
+// import data
+import { showMarkersLogic } from './showMarkersLogic.js'
 // requesting data from the user
 // getting the user location
 if (navigator.geolocation) {
@@ -13,27 +15,10 @@ if (navigator.geolocation) {
       // show user Location
 
       var map = L.map('map').setView([latitude, longitude], 13)
+      let userPosationText =
+        'Your current location <br> (this may not be your ex....)'
 
-      // my location
-      // 52.3624756,5.2083721,21
-
-      L.tileLayer('https://tile.openstreetmap.org/{z}/{x}/{y}.png', {
-        attribution:
-          '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors',
-      }).addTo(map)
-
-      var redIcon = L.icon({
-        iconUrl:
-          'https://raw.githubusercontent.com/pointhi/leaflet-color-markers/master/img/marker-icon-red.png',
-        iconSize: [25, 41],
-        iconAnchor: [12, 41],
-        popupAnchor: [1, -34],
-      })
-
-      L.marker([latitude, longitude], { icon: redIcon })
-        .addTo(map)
-        .bindPopup('Your current location <br> (this may not be your ex....)')
-        .openPopup()
+      showMarkersLogic(map, latitude, longitude, userPosationText)
     },
     () => {
       console.log('There is no location')
